% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{add_counters}
\alias{add_counters}
\title{Add externally saved counters}
\usage{
add_counters(cnt)
}
\arguments{
\item{cnt}{Counters to add, in an environment.}
}
\description{
Add externally saved counters
}
