% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-files.R
\name{file_coverage}
\alias{file_coverage}
\alias{format.files_coverage}
\alias{format.file_coverage}
\alias{print.file_coverage}
\alias{print.files_coverage}
\title{Show code coverage for one or more files, line by line}
\usage{
file_coverage(cov, include_paths = NULL)

\method{format}{files_coverage}(x, ...)

\method{format}{file_coverage}(x, ...)

\method{print}{file_coverage}(x, ...)

\method{print}{files_coverage}(x, pager = interactive() && length(sys.calls()) <= 2, ...)
}
\arguments{
\item{cov}{Code coverage result, typically from the \code{\link[=parse_coverage]{parse_coverage()}}
function.}

\item{include_paths}{If not \code{NULL}, then it may be a character vector of
paths, to only include files within these paths.}

\item{x}{\code{file_coverage} or \code{files_coverage} object.}

\item{...}{Currently ignored.}

\item{pager}{Whether to show the result in the system pager.
The default tries to guess when this makes sense. It also tries to
guess whether the default pager can show multiple files.}
}
\value{
A list with class \code{files_coverage}. Each list element is a
character vector with attibutes \code{filename}, \code{total}, \code{covered},
\code{coverage} and class \code{file_coverage}.

Both \code{files_coverage} and \code{file_coverage} have \code{format()} and
\code{print()} methods.
}
\description{
Show code coverage for one or more files, line by line
}
\seealso{
Other code coverage reporting functions: 
\code{\link{summary.code_coverage}()},
\code{\link{uncovered_lines}()}
}
\concept{code coverage reporting functions}
