% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-functions.R
\name{find_functions}
\alias{find_functions}
\title{Find all functions in an environment or package}
\usage{
find_functions(env)
}
\arguments{
\item{env}{An environment or package name to find functions in.}
}
\description{
Find all functions in an environment or package
}
