% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore_coverage}
\alias{ignore_coverage}
\title{Remove ignored code lines from code coverage data}
\usage{
ignore_coverage(cov, root = ".", ignore_files = NULL)
}
\arguments{
\item{cov}{Code coverage object, typically from the
\code{\link[=parse_coverage]{parse_coverage()}} function.}

\item{root}{Project root, the globs in \code{ignore_files} are interpreted
relative to this directory.}

\item{ignore_files}{Character vector of globs}
}
\description{
It ignores
}
\details{
\itemize{
\item all files matching the globs in \code{ignore_files}.
\item all lines between a \verb{# nocov start} and \verb{# nocov end} comment.
\item all lines that have a \verb{# nocov} comment.
}

It warns for unpaired \verb{# nocov start} and \verb{# nocov end} comments.
}
