% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-coverage.R
\name{parse_coverage}
\alias{parse_coverage}
\title{Summarize and parse all test coverage}
\usage{
parse_coverage(root = ".")
}
\arguments{
\item{root}{Root path, typically the root of an R package tree.}
}
\value{
Data frame with columns:
\itemize{
\item \code{file}: Absolute path to source file.
\item \code{line}: Line number, numbering start with 1.
\item \code{coverage}: Number of times this line was executed. If \code{NA}, then
it is not a code line that can be executed.
\item \code{code}: Character vector, the actual code.
}
}
\description{
The result includes both the R and C test coverage output.
}
