% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-coverage.R
\name{parse_covr_file}
\alias{parse_covr_file}
\title{Parse an R coverage output file}
\usage{
parse_covr_file(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
Data frame with columns:
\itemize{
\item \code{file}: Absolute path to source file.
\item \code{line}: Line number, numbering starts with 1.
\item \code{coverage}: Number of times this line was executed. If \code{NA}, then
it is not a code line that can be executed.
\item \code{code}: Character vector, the actual code.
}
}
\description{
Parse an R coverage output file
}
\seealso{
Other coverage output parser functions: 
\code{\link{parse_r_coverage}()}
}
\concept{coverage output parser functions}
