% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcov.R
\name{parse_gcov}
\alias{parse_gcov}
\title{Parse all \code{.gcov} files within a directory}
\usage{
parse_gcov(root = ".")
}
\arguments{
\item{root}{Root path, typically the root of an R package tree.}
}
\value{
Data frame with columns:
\itemize{
\item \code{file}: Absolute path to source file.
\item \code{line}: Line number, numbering start with 1.
\item \code{coverage}: Number of times this line was executed. If \code{NA}, then
it is not a code line that can be executed.
\item \code{code}: Character vector, the actual code.
}
}
\description{
Finds, parses and summarizes all \code{.gcov} files with code coverage
information, within a root directory.
}
\seealso{
Other gcov functions: 
\code{\link{parse_gcov_file}()},
\code{\link{run_gcov}()}
}
\concept{gcov functions}
