% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcov.R
\name{parse_gcov_file}
\alias{parse_gcov_file}
\title{Parse a \code{.gcov} file}
\usage{
parse_gcov_file(path)
}
\arguments{
\item{path}{Path to the \code{.gcov} file.}
}
\value{
Data frame with columns:
\itemize{
\item \code{file}: Absolute path to source file.
\item \code{line}: Line number, numbering start with 1.
\item \code{coverage}: Number of times this line was executed. If \code{NA}, then
it is not a code line that can be executed.
\item \code{code}: Character vector, the actual code.
}
}
\description{
Parse a \code{.gcov} file
}
\seealso{
Other gcov functions: 
\code{\link{parse_gcov}()},
\code{\link{run_gcov}()}
}
\concept{gcov functions}
