% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-coverage.R
\name{parse_r_coverage}
\alias{parse_r_coverage}
\title{Parse all coverage output or R files}
\usage{
parse_r_coverage(root = ".")
}
\arguments{
\item{root}{Root path, typically root of an R package tree.}
}
\value{
Data frame with columns:
\itemize{
\item \code{file}: Absolute path to source file.
\item \code{line}: Line number, numbering starts with 1.
\item \code{coverage}: Number of times this line was executed. If \code{NA}, then
it is not a code line that can be executed.
\item \code{code}: Character vector, the actual code.
}
}
\description{
Parse all coverage output or R files
}
\seealso{
Other coverage output parser functions: 
\code{\link{parse_covr_file}()}
}
\concept{coverage output parser functions}
