% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcov.R
\name{run_gcov}
\alias{run_gcov}
\title{Run \code{gcov} summarize code coverage}
\usage{
run_gcov(root = ".")
}
\arguments{
\item{root}{Root path, typically the root of an R package tree.}
}
\description{
Calls \code{gcov} to create \code{.gcov} files for all \code{.gcno} files, which were
typically created when running a program instrumented with code
coverage.
}
\details{
It summarized all \code{.gcno} files, recursively.
}
\seealso{
Other gcov functions: 
\code{\link{parse_gcov_file}()},
\code{\link{parse_gcov}()}
}
\concept{gcov functions}
