% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{save_env_to_raw}
\alias{save_env_to_raw}
\title{Serialize an environment into a raw vector}
\usage{
save_env_to_raw(env, skip = c(".__NAMESPACE__."), closxp_callback = NULL)
}
\arguments{
\item{env}{Enviropnment to serialize.}

\item{skip}{Names of objects in \code{env} to skip.}

\item{closxp_callback}{If not \code{NULL}, then it must be a callback
function to transform closures. It is called for every function. It is
also called on functions embedded into functions. (First on the outer
function, and if it returns a function that embeds functions, then
on those.)}
}
\description{
This function is even more experimental than this experimental package.
}
\seealso{
\code{\link[base:save]{base::save()}}
}
