% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{save_rds}
\alias{save_rds}
\title{Serialize an R object into an \code{.rds} file}
\usage{
save_rds(object, file, closxp_callback = NULL)
}
\arguments{
\item{object}{R object to serialize.}

\item{file}{Output file.}

\item{closxp_callback}{If not \code{NULL}, then it must be a callback
function to transform closures. It is called for every function. It is
also called on functions embedded into functions. (First on the outer
function, and if it returns a function that embeds functions, then
on those.)}
}
\description{
This function is similar to \code{\link[base:readRDS]{base::saveRDS()}}, but it uses a
different serialization implementation. Differences:
}
\details{
\itemize{
\item Slightly faster implementation, in most cases.
\item Multi-byte output is in host byte-order (the same as choosing
\code{xdr = FALSE} for \code{\link[base:serialize]{base::serialize()}}).
\item No compression.
\item It always uses workspace format 2, and expands ALTREP objects.
\item Has a hook for closures, the \code{closxp_callback} argument.
\item Does not have a hook for reference objects.
}
}
\seealso{
\code{\link[=serialize_to_raw]{serialize_to_raw()}}, \code{\link[base:readRDS]{base::saveRDS()}}, \code{\link[base:readRDS]{base::readRDS()}}.
}
