% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{serialize_to_raw}
\alias{serialize_to_raw}
\title{Serialize R objects}
\usage{
serialize_to_raw(object, closxp_callback = NULL)
}
\arguments{
\item{object}{Object to serialize.}

\item{closxp_callback}{If not \code{NULL}, then it must be a callback
function to transform closures. It is called for every function. It is
also called on functions embedded into functions. (First on the outer
function, and if it returns a function that embeds functions, then
on those.)}
}
\value{
Raw vector, serialization of \code{object}.
}
\description{
Equivalent to base R' serialization format, with parameters
\code{connection = NULL}, \code{ascii = FALSE}, \code{xdr = FALSE}, \code{version = 2},
\code{refhook = NULL}.
}
\seealso{
\code{\link[base:serialize]{base::serialize()}}, \code{\link[base:serialize]{base::unserialize()}}.
}
