% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summary.code_coverage}
\alias{summary.code_coverage}
\title{Summarize code coverage results}
\usage{
\method{summary}{code_coverage}(object, paths = NULL, ...)
}
\arguments{
\item{object}{Code coverage object, typically from the
\code{\link[=parse_coverage]{parse_coverage()}} function.}

\item{paths}{If not \code{NULL}, then it may be a character vector of
paths, to only include files within these paths.}

\item{...}{Currently ignored.}
}
\value{
Data frame with columns:
\itemize{
\item \code{path}: absolute file path.
\item \code{relpath}: relative file path, calculated by dropping the common
prefix.
\item \code{total}: total number of code lines.
\item \code{coveraged}: number of covered lines.
\item \code{coverage}: code coverage ratio for this this.
\item \code{uncovered}: uncovered lines, see \code{\link[=uncovered_lines]{uncovered_lines()}} for the
format.
}
}
\description{
Summarize code coverage results
}
\seealso{
Other code coverage reporting functions: 
\code{\link{uncovered_lines}()}
}
\concept{code coverage reporting functions}
