% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{trace_calls}
\alias{trace_calls}
\title{Trace calls in a function}
\usage{
trace_calls(x, compile = TRUE)
}
\arguments{
\item{x}{Function to trace.}

\item{compile}{Logical scalar, whether to byte-compile the traced
function.}
}
\value{
The traced function
}
\description{
Trace calls in a function
}
