% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{trace_package}
\alias{trace_package}
\title{Trace functions in a package}
\usage{
trace_package(package)
}
\arguments{
\item{package}{Package name.}
}
\description{
Uses \code{\link[=transform_package]{transform_package()}} with \code{\link[=trace_calls]{trace_calls()}} as the callback for
function objects.
}
