% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_env}
\alias{transform_env}
\title{Transform all objects in an environment}
\usage{
transform_env(env, skip = NULL, closxp_callback = NULL)
}
\arguments{
\item{env}{Environment to transform}

\item{skip}{Character vector of object names to skip.}

\item{closxp_callback}{If not \code{NULL}, then it must be a callback
fucntion to transform closures. It is called for every function, but
it is not called for functions nested into other functions.}
}
\description{
This function is experimental.
}
\seealso{
Other transform functions: 
\code{\link{transform_object}()},
\code{\link{transform_package}()}
}
\concept{transform functions}
