% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_object}
\alias{transform_object}
\title{Transform an object}
\usage{
transform_object(object, closxp_callback = NULL)
}
\arguments{
\item{object}{R object to transform.}

\item{closxp_callback}{If not \code{NULL}, then it must be a callback
fucntion to transform closures. It is called for every function, but
it is not called for functions nested into other functions.}
}
\value{
The transformed object.
}
\description{
This function is experimental.
}
\seealso{
Other transform functions: 
\code{\link{transform_env}()},
\code{\link{transform_package}()}
}
\concept{transform functions}
