% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{uncovered_lines}
\alias{uncovered_lines}
\title{Report lines \emph{not} covered by tests, for each file}
\usage{
uncovered_lines(cov)
}
\arguments{
\item{cov}{Code coverage object, typically from the
\code{\link[=parse_coverage]{parse_coverage()}} function.}
}
\value{
Data frame with columns:
\itemize{
\item \code{path}: full absolute path.
\item \code{uncovered}: Uncovered lines, in a matrix with two rows. Each
column is an uncovered interval, the first row number has the
beginnings, the second the ends of the intervals.
}
}
\description{
Report lines \emph{not} covered by tests, for each file
}
\seealso{
Other code coverage reporting functions: 
\code{\link{file_coverage}()},
\code{\link{summary.code_coverage}()}
}
\concept{code coverage reporting functions}
